+++ ha-api/src/main/java/jp/co/ha/api/response/HealthInfoRegistResponse.java
@@ -12,7 +12,7 @@ import jp.co.ha.common.log.annotation.Mask;
 public class HealthInfoRegistResponse extends BaseResponse {
 
 	/** 健康情報ID */
-	private BigDecimal healthInfoId;
+	private Integer healthInfoId;
 	/** ユーザID */
 	private String userId;
 	/** 身長 */
@@ -37,7 +37,7 @@ public class HealthInfoRegistResponse extends BaseResponse {
 	 *
 	 * @return healthInfoId 健康情報ID
 	 */
-	public BigDecimal getHealthInfoId() {
+	public Integer getHealthInfoId() {
 		return healthInfoId;
 	}
 
@@ -47,7 +47,7 @@ public class HealthInfoRegistResponse extends BaseResponse {
 	 * @param healthInfoId
 	 *     健康情報ID
 	 */
-	public void setHealthInfoId(BigDecimal healthInfoId) {
+	public void setHealthInfoId(Integer healthInfoId) {
 		this.healthInfoId = healthInfoId;
 	}
